+++ src/builtin.h
@@ -1822,9 +1822,9 @@ ARSH_METHOD array_sort(RuntimeContext &ctx) {
   RET(LOCAL(0));
 }
 
-//!bind: function sortWith($this : Array<T0>, $comp : Func<Bool, [T0, T0]>) : Array<T0>
-ARSH_METHOD array_sortWith(RuntimeContext &ctx) {
-  SUPPRESS_WARNING(array_sortWith);
+//!bind: function sortBy($this : Array<T0>, $comp : Func<Bool, [T0, T0]>) : Array<T0>
+ARSH_METHOD array_sortBy(RuntimeContext &ctx) {
+  SUPPRESS_WARNING(array_sortBy);
   auto &obj = typeAs<ArrayObject>(LOCAL(0));
   CHECK_ITER_INVALIDATION(obj);
   auto &comp = LOCAL(1);