+++ test/api/api_test.cpp
@@ -474,7 +474,7 @@ TEST_F(APITest, abort) {
   ASSERT_EQ(1, s);
   ASSERT_EQ(AR_ERROR_KIND_TYPE_ERROR, e->kind);
 
-  src = R"(["d", ""].sortWith($f))";
+  src = R"(["d", ""].sortBy($f))";
   e = newError();
   s = ARState_eval(this->state, nullptr, src.c_str(), src.size(), e.get());
   ASSERT_EQ(1, s);
@@ -486,7 +486,7 @@ TEST_F(APITest, abort) {
   ASSERT_EQ(0, s);
   ASSERT_EQ(AR_ERROR_KIND_SUCCESS, e->kind);
 
-  src = R"(["d", ""].sortWith($g))";
+  src = R"(["d", ""].sortBy($g))";
   e = newError();
   s = ARState_eval(this->state, nullptr, src.c_str(), src.size(), e.get());
   ASSERT_EQ(54, s);