+++ CHANGELOG.md
@@ -52,6 +52,7 @@
 #### Builtin
 
 - **Breaking Change**: ``String#sanitize`` method no longer replace null characters
+- **Breaking Change**: rename ``Array#sortWith`` method with ``Array#sortBy``
 - now interrupt ``LineEditor#readLine`` method when receive signals
     - also call signal handler
 - natively complete builtin `shctl` command sub-commands/options. now no longer need completion script